Implemented basic 'tag' feature.
This feature lets you give tougou-chan a tag and a coresponding word to remember,
She can then recall the word given the tag or delete tags.
She saves tags to a file on the system which is in .gitignore. this means users can create their own tags which are stored between restarts/updates.