add some testing!
One of my intentions for this rewrite is to greatly improve the testhing
behind tougou.
I was originally very naive and didn't think testing made sense in this
application. my opinion on this has very much changed and I'm greatly
looking forward to having a great suite of automated tests!